api/schemas: Move "geometry" field into "skylines.schemas.fields"
