+++ skylines/schemas/fields.py
@@ -1,5 +1,9 @@
+from marshmallow import ValidationError
 from marshmallow.fields import *  # NOQA
 from marshmallow.fields import String as _String
+
+from geoalchemy2.shape import to_shape
+from shapely.geometry import Polygon, Point
 
 
 class String(_String):
@@ -19,6 +23,36 @@ class String(_String):
 class Location(Field):
     def _serialize(self, value, attr, obj):
         return value.to_lonlat() if value else None
+
+
+class GeometryField(Field):
+    def _serialize(self, value, attr, obj):
+        if value is None:
+            return None
+
+        shape = to_shape(value)
+        if isinstance(shape, Point):
+            return self.serialize_point(shape)
+        if isinstance(shape, Polygon):
+            return self.serialize_polygon(shape)
+
+        raise ValidationError('Unsupported shape type')
+
+    @classmethod
+    def serialize_polygon(cls, polygon):
+        exterior = polygon.exterior
+        if exterior is None:
+            return None
+
+        return map(cls.serialize_coord, exterior.coords)
+
+    @classmethod
+    def serialize_point(cls, point):
+        return cls.serialize_coord(point.coords[0])
+
+    @staticmethod
+    def serialize_coord(coord):
+        return [coord[0], coord[1]]
 
 
 # Aliases
